% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam2formula.R
\name{gam2formula}
\alias{gam2formula}
\title{Derive formulas for all B-spline, P-spline and cubic regression spline smooths of a
generalized additive model fitted with \code{mgcv}.}
\usage{
gam2formula(mod)
}
\arguments{
\item{mod}{A fitted model object of class \code{gam} or \code{bam} with at least 
one smooth of type \code{"bs"} (B-spline), \code{"ps"} (P-spline) or 
\code{"cr"} (cubic regression spline).}
}
\value{
An object of class \code{gam2formula}. This is a nested tibble with one 
row per supported smooth in the model object. The derived formulas are stored in
the column \code{coeftab}.
}
\description{
Derive formulas for all B-spline, P-spline and cubic regression spline smooths of a
generalized additive model fitted with \code{mgcv}.
}
\details{
This function derives formulas for all B-spline, P-spline and cubic regression
spline smooths of a generalized additive model fitted with \code{\link[mgcv:gam]{gam()}} 
or \code{\link[mgcv:bam]{bam()}}. To this end, the function uses a simple basis 
representation for each supported smooth and computes the basis coefficients 
using ordinary least squares. Predictions from the derived formulas are 
typically in perfect agreement with predictions 
from \code{\link[mgcv:predict.gam]{predict.gam()}} up to numerical error 
of negligible magnitude. The function works for any type of 
response including binary, continuous, count or time-to-event responses.
}
\section{Warnings}{
 There is no
support for thin plate regression splines of class \code{"tp"} 
(the default in \code{mgcv}), as these do not have a simple basis representation. 
In addition, there is currently no support for smooths that contain 
a \code{by = } statement, multivariate smooths of 
classes \code{"te"} or \code{"ti"} or fits from
\code{\link[mgcv:bam]{bam()}} with option \code{discrete = TRUE}.
}

\examples{
# generate data and fit gam with mgcv
library(mgcv)
set.seed(2)
dat <- gamSim(1, n = 200, dist = "normal", scale = 2, verbose = FALSE)
mod <- gam(y ~ s(x0, bs = "cr") + s(x1, bs = "bs", m = c(4, 2)) +
s(x2, bs = "ps") + s(x3), data = dat)

# derive closed formulas for all supported smooths using gam2formula
mod_formulas <- gam2formula(mod)

# get an overview of all derived formulas
summary(mod_formulas)

# print derived formula for s(x0)
print(mod_formulas, term = "x0")

# print derived formula in LaTeX format
print(mod_formulas, term = "x0", format = "latex")

# predict from the derived formula of s(x0) at original data points
predict(mod_formulas, term = "x0", newdata = dat)

# predict from the derived formula at new data points
predict(mod_formulas, term = "x0", 
newdata = data.frame(x0 = seq(min(dat$x0), max(dat$x0), len = 100)))

}
\seealso{
Use the function \code{\link[=summary.gam2formula]{summary()}} 
for an overview, the function \code{\link[=print.gam2formula]{print()}} to print 
the derived formula of a smooth, and the function 
\code{\link[=predict.gam2formula]{predict()}} to predict from the derived 
formula of a smooth.
}
