% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gam2formula}
\alias{print.gam2formula}
\title{Print the derived formula of a smooth.}
\usage{
\method{print}{gam2formula}(x, term = NULL, format = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{gam2formula} as returned by 
\code{\link[=gam2formula]{gam2formula()}}.}

\item{term}{Can either be \code{NULL} (default) or a string with the name of 
the predictor for which to print
the formula (must be included in \code{x} and can only be of length one).}

\item{format}{Can either be \code{NULL} (default) or \code{"latex"}.}

\item{...}{Other arguments.}
}
\description{
The default printing method for an object of class \code{gam2formula}.
}
\details{
When \code{term = NULL}, an overview is printed via
 \code{\link[=summary.gam2formula]{summary()}}.
When \code{term} is given with \code{format = NULL}, the coefficient 
table is returned as a tibble.
When \code{format = "latex"}, LaTeX code is printed.
}
\examples{
# generate data and fit gam with mgcv
library(mgcv)
set.seed(2)
dat <- gamSim(1, n = 200, dist = "normal", scale = 2, verbose = FALSE)
mod <- gam(y ~ s(x0, bs = "cr") + s(x1, bs = "bs", m = c(4, 2)) +
s(x2, bs = "ps") + s(x3), data = dat)

# derive closed formulas for all supported smooths using gam2formula
mod_formulas <- gam2formula(mod)

# print derived formula for s(x0)
print(mod_formulas, term = "x0")

# print derived formula in LaTeX format
print(mod_formulas, term = "x0", format = "latex")

}
\seealso{
Use the function \code{\link[=summary.gam2formula]{summary()}} 
for an overview and the function  \code{\link[=predict.gam2formula]{predict()}} 
to predict from the derived formula of a smooth.
}
