% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{RBF_design}
\alias{RBF_design}
\title{Return the design matrix of the rescaled cubic radial basis function basis.}
\usage{
RBF_design(x, knots)
}
\arguments{
\item{x}{A vector of x_i values.}

\item{knots}{The vector of spline knots.}
}
\value{
The design matrix B=B_j(x_i) with intercept, linear, and cubic RBF terms.
}
\description{
Return the design matrix of the rescaled cubic radial basis function basis.
}
