% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth2formula_bs.R
\name{smooth2formula_bs}
\alias{smooth2formula_bs}
\title{Derive closed formula for a B-spline smooth}
\usage{
smooth2formula_bs(smooth, mod, ngrid = 1000)
}
\arguments{
\item{smooth}{A \code{"bs"} (B-spline) smooth.}

\item{mod}{The model object that the smooth forms part of.}

\item{ngrid}{The number of grid points on which to evaluate the smooth.}
}
\value{
A nested tibble with name of the term, type of the smooth (\code{"bs"}) and a closed formula representation of the smooth.
}
\description{
Derive closed formula for a B-spline smooth
}
