% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth2formula_cr.R
\name{smooth2formula_cr}
\alias{smooth2formula_cr}
\title{Derive closed formula for a cubic regression spline smooth}
\usage{
smooth2formula_cr(smooth, mod, ngrid = 1000)
}
\arguments{
\item{smooth}{A  \code{"cr"} (cubic regression spline) smooth.}

\item{mod}{The model object that the smooth forms part of.}

\item{ngrid}{The number of grid points on which to evaluate the smooth.}
}
\value{
A nested tibble with name of the term, type of the smooth (\code{"cr"}) and a closed formula representation of the smooth.
}
\description{
Derive closed formula for a cubic regression spline smooth
}
