% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_latex.R
\name{str_to_latex}
\alias{str_to_latex}
\title{Convert a string with an extracted spline formula into LaTeX syntax}
\usage{
str_to_latex(s, term, application)
}
\arguments{
\item{s}{The string containing the formula to be converted.}

\item{term}{The name of the term to be displayed as text, supplied as a string.}

\item{application}{Whether basis function formula (\code{"bfun"}) or
range formula (\code{"range"}) is converted.}
}
\description{
Convert a string with an extracted spline formula into LaTeX syntax
}
