% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.gam2formula}
\alias{summary.gam2formula}
\title{Summarize a gam2formula object.}
\usage{
\method{summary}{gam2formula}(object, ...)
}
\arguments{
\item{object}{An object of class \code{gam2formula} as returned by 
\code{\link[=gam2formula]{gam2formula()}}.}

\item{...}{Other arguments.}
}
\value{
The input object, invisibly.
}
\description{
The default summary method for an object of class \code{gam2formula}.
}
\examples{
# generate data and fit gam with mgcv
library(mgcv)
set.seed(2)
dat <- gamSim(1, n = 200, dist = "normal", scale = 2, verbose = FALSE)
mod <- gam(y ~ s(x0, bs = "cr") + s(x1, bs = "bs", m = c(4, 2)) +
s(x2, bs = "ps") + s(x3), data = dat)
mod_formulas <- gam2formula(mod)
summary(mod_formulas)

}
\seealso{
Use the function  \code{\link[=print.gam2formula]{print()}} to print 
the derived formula of a smooth and the function 
\code{\link[=predict.gam2formula]{predict()}} to predict from the derived 
formula of a smooth.
}
