% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{T_spline_design}
\alias{T_spline_design}
\title{Return the design matrix of the rescaled truncated power series basis.}
\usage{
T_spline_design(x, m, intknots, a, b)
}
\arguments{
\item{x}{A vector of x_i values.}

\item{m}{The degree of the spline.}

\item{intknots}{The vector of interior spline knots.}

\item{a}{Left boundary of spline interval.}

\item{b}{Right boundary of spline interval.}
}
\value{
The design matrix T=T_j(x_i).
}
\description{
Return the design matrix of the rescaled truncated power series basis.
}
