% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_latex.R
\name{write_latex}
\alias{write_latex}
\title{Write LaTeX code for coefficient table of derived formulas}
\usage{
write_latex(x, term)
}
\arguments{
\item{x}{An object of class \code{gam2formula} as returned by \code{\link[=gam2formula]{gam2formula()}}.}

\item{term}{A string with the name of the predictor for which to write LaTeX code.}
}
\value{
A string with LaTeX syntax for the coefficient table created with \code{kableExtra::kable}.
By using \code{cat} on it, it gets parsed in the R console
for copy-and-pasting into a LaTeX compiler.
}
\description{
Write LaTeX code for coefficient table of derived formulas
}
\details{
This function writes LaTeX code for derived formulas of B-spline, P-spline and cubic regression
spline smooths, relying on \code{kableExtra} for LaTeX code generation.
}
